backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pillow==10.4.0
opencv-python==4.10.0.84
mediapipe==0.10.14
rembg==2.0.57
numpy==1.26.4
